Improve dev experience when dealing with the tests
