+++ docker-compose.dev.yml
@@ -6,4 +6,4 @@ services:
     hostname: dynamodb-local
     container_name: dynamodb-local
     ports:
-      - 8042:8000
+      - 18000:8000
